+++ profile/src/main/java/com/scottlogic/deg/profile/reader/file/CsvInputStreamReader.java
@@ -44,7 +44,7 @@ public final class CsvInputStreamReader {
     }
 
     private static WeightedElement<String> createWeightedElement(CSVRecord record) {
-        final float weight = record.size() > 1 ? Float.parseFloat(record.get(1)) : 1.0F;
+        final double weight = record.size() > 1 ? Double.parseDouble(record.get(1)) : 1.0F;
         return new WeightedElement<>(record.get(0), weight);
     }
